+++ QMFAAuthVerifyAtDevice.html
@@ -18,7 +18,7 @@ WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN 
     <!-- Bootstrap CSS -->
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
     <link rel="stylesheet" type="text/css" href="./css/site.css"/>
-    <script type="module" src="./js/qmfaAuthatdevice.js"></script>
+    <script type="module" src="./js/Demo_qmfaAuthatdevice.js"></script>
   </head>
   <body>
     <div class="container">
