+++ index.html
@@ -67,7 +67,7 @@ WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN 
 		</div>
 		<footer class="mt-auto text-center text-lg-start text-muted fixed-bottom">
 			<div class="mt-auto text-center p-4" style="background-color: rgba(0, 0, 0, 0);">
-			  MagTek, Inc. Copyright 2024 ©: Version 1.0.0.3
+			  MagTek, Inc. Copyright 2024 ©: Version 1.1.0.0
 			</div>
 		</footer>
 	</body>
