+++ js/API_rms.js
@@ -11,9 +11,6 @@ DO NOT REMOVE THIS COPYRIGHT
 */
 
 "use strict";
-//import * as mt_V5 from "./mt_v5.js";
-//import * as mt_Utils from "./mt_utils.js";
-
 
 export let BaseURL = "";
 export let APIKey =  "";
