+++ js/appsettings.js
@@ -2,5 +2,4 @@ export const MQTT  = {
     Base_Pub : "MagTek/Server/",
     Base_Sub : "MagTek/Device/",
     DeviceList : "MagTek/Server/+/+/Status"
-}
-
+};
