+++ mmsmqttdevice.html
@@ -44,7 +44,8 @@ You can minimize this page without affecting the connection, but it must remain 
         <button class="btn btn-primary" id="deviceClose">Close</button>      
       </div>
       <hr>
-      <img id= "QRCode" src="">
+      <a id= QRCodeLink href="" target="_blank"><img id= "QRCode" src=""></a>
+      
       <div class="container">
         <div class="form-group row">
           <div class="col-md">
